Generate only kpis

+++ chronix2grid/main.py
@@ -82,6 +82,18 @@ def generate_mp_core(prng, case, start_date, weeks, by_n_weeks, n_scenarios, mod
         case, start_date, output_folder, scenario_base_name, n_scenarios, mode,
         warn_user=not ignore_warnings)
 
+    ###if only 'K' mode, the genreation output folder can already be the output folder, check that. but inform that it is not the conventional one
+    if mode=='K':
+        print("as the time series as already been generated, checking if they are located in "+generation_output_folder+" or in root folder "
+              +output_folder)
+        if os.path.exists(generation_output_folder) and len(os.listdir(generation_output_folder))!=0:
+            print("the data time series folder is identified as "+generation_output_folder)
+        else:
+            generation_output_folder=output_folder
+            kpi_output_folder = os.path.join(output_folder, cst.KPI_FOLDER_NAME)
+    ##########
+
+
     # seeds
     default_seed = generate_default_seed(prng)
     seed_for_loads = parse_seed_arg(seed_for_loads, '--seed-for-loads',
@@ -99,7 +111,8 @@ def generate_mp_core(prng, case, start_date, weeks, by_n_weeks, n_scenarios, mod
 
     print('initial_seeds')
     print(initial_seeds)
-    dump_seeds(generation_output_folder, initial_seeds, scenario_name)
+    if mode!='K':#if data is being generated, so if we are not only computing KPIs on existing generated data
+        dump_seeds(generation_output_folder, initial_seeds, scenario_name)
 
     if n_scenarios >= 2:
         seeds_for_loads, seeds_for_res, seeds_for_disp = gu.generate_seeds(
@@ -111,26 +124,30 @@ def generate_mp_core(prng, case, start_date, weeks, by_n_weeks, n_scenarios, mod
         seeds_for_loads = [seed_for_loads]
         seeds_for_res = [seed_for_res]
         seeds_for_disp = [seed_for_dispatch]
-
+    if nb_core==1:
+        for i in range(n_scenarios):
+            generate_per_scenario(case, start_date, weeks, by_n_weeks, mode, input_folder,
+            kpi_output_folder, generation_output_folder, scen_names,
+            seeds_for_loads, seeds_for_res, seeds_for_disp, ignore_warnings,i)
+    else:
     # multi-processing
-    pool = multiprocessing.Pool(nb_core)
-    iterable = [i for i in range(n_scenarios)]
-    multiprocessing_func = partial(
-        generate_per_scenario,
-        case, start_date, weeks, by_n_weeks, mode, input_folder,
-        kpi_output_folder, generation_output_folder, scen_names,
-        seeds_for_loads, seeds_for_res, seeds_for_disp, ignore_warnings)
+        with multiprocessing.Pool(nb_core) as pool:
+            iterable = [i for i in range(n_scenarios)]
+            multiprocessing_func = partial(
+                generate_per_scenario,
+                case, start_date, weeks, by_n_weeks, mode, input_folder,
+                kpi_output_folder, generation_output_folder, scen_names,
+                seeds_for_loads, seeds_for_res, seeds_for_disp, ignore_warnings)
 
-    pool.map(multiprocessing_func, iterable)
-    pool.close()
-    print('multiprocessing done')
-    print('Time taken = {} seconds'.format(time.time() - start_time))
-    print('removing temporary folders if exist:')
-    rm_temporary_folders(input_folder, case)
+            pool.map(multiprocessing_func, iterable)
+        print('multiprocessing done')
+        print('Time taken = {} seconds'.format(time.time() - start_time))
+        print('removing temporary folders if exist:')
+    rm_temporary_folders(input_folder,output_folder, case)
 
-def rm_temporary_folders(input_folder, case):
+def rm_temporary_folders(input_folder,output_folder, case):
     grid2op_tempo = os.path.join(input_folder, cst.GENERATION_FOLDER_NAME, case, 'chronics')
-    if os.path.exists(grid2op_tempo):
+    if os.path.exists(grid2op_tempo) and grid2op_tempo!=output_folder:
         shutil.rmtree(grid2op_tempo)
         print("--"+str(grid2op_tempo)+" deleted")
 
@@ -162,7 +179,9 @@ def generate_per_scenario(case, start_date, weeks, by_n_weeks, mode,
 
     scenario_path = os.path.join(generation_output_folder, scenario_name)
     print('scenario_path: '+scenario_path)
-    dump_seeds(scenario_path, scenario_seeds)
+
+    if mode != 'K':  # if data is generated
+        dump_seeds(scenario_path, scenario_seeds)
 
     # go to generate chronics
     generate_inner(
@@ -188,13 +207,14 @@ def generate_inner(case, start_date, weeks, by_n_weeks, n_scenarios, mode,
 
     year = time_parameters['year']
 
+
     # Chronic generation
     if 'L' in mode or 'R' in mode:
         generator = GeneratorBackend()
-        params, loads_charac, prods_charac = gen.main(generator,
-            case, n_scenarios, generation_input_folder,
-            generation_output_folder, scen_names, time_parameters,
-            mode, scenario_id, seed_for_loads, seed_for_res, seed_for_dispatch)
+        params, _, _ = gen.main(generator,case, n_scenarios, generation_input_folder,
+                                 generation_output_folder, scen_names, time_parameters,
+                                 mode, scenario_id, seed_for_loads, seed_for_res,
+                                 seed_for_dispatch)
         scenario_name = scen_names(scenario_id)
         if by_n_weeks is not None and 'T' in mode:
             output_processor_to_chunks(
@@ -205,19 +225,18 @@ def generate_inner(case, start_date, weeks, by_n_weeks, n_scenarios, mode,
                 n_scenarios, start_date, int(params['dt']))
 
     # KPI formatting and computing
-    if 'R' in mode and 'K' in mode and 'T' not in mode:
-        # Get and format solar and wind on all timescale, then compute KPI and save plots
-        wind_solar_only = True
-        kpis.main(kpi_input_folder, generation_output_folder, scen_names,
-                  kpi_output_folder, year, case, n_scenarios, wind_solar_only,
-                  params, loads_charac, prods_charac, scenario_id)
+    if 'K' in mode :
+        #get params, prods_charac and loads_charac from Generator config
+        generator = GeneratorBackend()
+        config_manager_dict = generator.get_config_managers(generation_input_folder, case, generation_output_folder, mode)
+        params_dict, prods_charac, loads_charac = generator.get_params_charact(time_parameters, config_manager_dict)
 
-    elif 'T' in mode and 'K' in mode:
-        # Get and format dispatched chronics, then compute KPI and save plots
-        wind_solar_only = False
+        print("WARNING: make sure that your generated data is in folder: "+generation_output_folder)
+
+        #compute KPIs
         kpis.main(kpi_input_folder, generation_output_folder, scen_names,
-                  kpi_output_folder, year, case, n_scenarios, wind_solar_only,
-                  params, loads_charac, prods_charac, scenario_id)
+                  kpi_output_folder, year, case, n_scenarios,
+                  params_dict['G'], loads_charac, prods_charac, scenario_id)
 
 
 def create_directory_tree(case, start_date, output_directory, scenario_name,
@@ -226,7 +245,8 @@ def create_directory_tree(case, start_date, output_directory, scenario_name,
         output_directory, cst.GENERATION_FOLDER_NAME, case, start_date)
     if warn_user and os.path.isdir(gen_path_to_create):
         gu.warn_if_output_folder_not_empty(gen_path_to_create)
-    os.makedirs(gen_path_to_create, exist_ok=True)
+    if(mode!='K'):#if time series have not already been generated
+        os.makedirs(gen_path_to_create, exist_ok=True)
 
     kpi_path_to_create = None
     if 'K' in mode:
@@ -241,13 +261,14 @@ def create_directory_tree(case, start_date, output_directory, scenario_name,
     for i in range(n_scenarios):
         s_name = scen_name_generator(i)
         scenario_path_to_create = os.path.join(gen_path_to_create, s_name)
-        os.makedirs(scenario_path_to_create, exist_ok=True)
+        if (mode != 'K'):  # if time series have not already been generated
+            os.makedirs(scenario_path_to_create, exist_ok=True)
 
-        if 'K' in mode:
-            scenario_kpi_path_to_create = os.path.join(
-                kpi_path_to_create, s_name, cst.KPI_IMAGES_FOLDER_NAME
-            )
-            os.makedirs(scenario_kpi_path_to_create, exist_ok=True)
+            if 'K' in mode:
+                scenario_kpi_path_to_create = os.path.join(
+                    kpi_path_to_create, s_name, cst.KPI_IMAGES_FOLDER_NAME
+                )
+                os.makedirs(scenario_kpi_path_to_create, exist_ok=True)
 
     return gen_path_to_create, kpi_path_to_create
